Allow Scheduler Simulator to "defer dispatch" until controlled
This is needed to prevent a semaphore delete from triggering a
preempt and breaking the mutex is owner debug check. Technically
we are running the same thread until the API call returns. But from
an RTEMS internals perspective, it was preempted at the _Object_Put().

When understanding this patch, it is important to remember that the
Scheduler Simulator is single threaded and faking out RTEMS. This
is one place where RTEMS would have preempted to a thread and when
we returned to rtems_semaphore_delete(), the _Object_Put() call
would have been made by the correct thread.